minor refactoring, more SOLID compliant

+++ parking_lot/models/strategy/fare_strategy.py
@@ -5,7 +5,7 @@ from parking_lot.constants import VehicleType
 
 
 class FareStrategy(ABC):
-    def calculate_fare(self, vehicle_type: VehicleType, duration_seconds: int) -> int:
+    def calculate_fare(self, vehicle_type: VehicleType, duration_seconds: int) -> float:
         pass
 
 class SimpleFareStrategy(FareStrategy):
@@ -17,7 +17,7 @@ class SimpleFareStrategy(FareStrategy):
         }
         self.default_fare = 20
 
-    def calculate_fare(self, vehicle_type: VehicleType, duration_seconds: int) -> int:
+    def calculate_fare(self, vehicle_type: VehicleType, duration_seconds: int) -> float:
         fare = math.ceil(duration_seconds/3600) * self.fare_map.get(vehicle_type, self.default_fare)
         print(f"[SimpleFareStrategy]: Fare for vehicle_type {vehicle_type.value} for duration {duration_seconds} is ${fare}.")
         return fare
